More defined list-to-V/M operators: (|>|), (>|<).
Provides analogs of the (|>) and (><) operators that first apply an
appropriate 'take' to the given lists so that they may be safely used
on lists that are too long (or infinite) -- a feature I find 
particularly useful from the interactive prompt.

As these operators are more defined, I would ask the package maintainer 
to consider, rather than adding (|>|) and (>|<), just updating the
definitions of (|>) and (><) with the new semantics.