Fix real and potential memory leaks in C code
Use an unwinding method to free memory in the event of failure.

There's three idioms at use here.

CHECK; MARK/CONVERGED; UNWIND.

CHECK is as it was, it exits the function immediately with a code.

MARK and CONVERGED set the variable `ret` to a suitable exit code
for the function, but don't return or jump.

UNWIND jumps to the point where it will free all the memory that
has been allocated so far.